requirements.txt
streamlit
spacy
stable-baselines3
gymnasium
pandas
tsfresh
google-generativeai

# We also need the model for spaCy

en_core_web_sm @ https://www.google.com/search?q=https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz